Reorganize extensions
