+++ src/Common/Common.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>Immediate.Validations</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ConstraintsExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ITypeSymbolExtensions.cs" />
   </ItemGroup>
